try to refactor the acl backend (#536)
